+++ mecha/_ellip.py
@@ -148,7 +148,7 @@ class EllipticSpirograph(DrawingMechanism):
             """Simulate one cycle of the assembly's motion."""
             length = self.get_cycle_length()
             if self.per_turn:
-                nb_samples = (length / (2*math.pi)) * self.nb_samples + 1
+                nb_samples = int(length / (2*math.pi)) * self.nb_samples + 1
             else:
                 nb_samples = self.nb_samples
             curve = self.roul.get_range(0., length, nb_samples, reuse)
